Refactor the tasks so that the in-proc code is in partial classes.
